JBPM-5530 - Forms modeler packages, artifact id and GWT module rename
 Update references to use workbench instead of console.ng
 Drop human-task from artifact id as it relates to both processes and tasks forms
 Update GWT modules name
