+++ jbpm-console-ng-showcase/pom.xml
@@ -334,13 +334,13 @@
 
         <dependency>
             <groupId>org.jbpm</groupId>
-            <artifactId>jbpm-wb-human-tasks-forms-modeler-client</artifactId>
+            <artifactId>jbpm-wb-forms-modeler-client</artifactId>
             <scope>provided</scope>
         </dependency>
 
         <dependency>
             <groupId>org.jbpm</groupId>
-            <artifactId>jbpm-wb-human-tasks-forms-modeler-backend</artifactId>
+            <artifactId>jbpm-wb-forms-modeler-backend</artifactId>
         </dependency>
 
         <dependency>
@@ -1294,9 +1294,9 @@
                         <compileSourcesArtifact>org.jbpm:jbpm-wb-human-tasks-client</compileSourcesArtifact>
                         <compileSourcesArtifact>org.jbpm:jbpm-wb-human-tasks-forms-api</compileSourcesArtifact>
                         <compileSourcesArtifact>org.jbpm:jbpm-wb-human-tasks-forms-client</compileSourcesArtifact>
-                        <compileSourcesArtifact>org.jbpm:jbpm-wb-human-tasks-forms-modeler-api</compileSourcesArtifact>
+                        <compileSourcesArtifact>org.jbpm:jbpm-wb-forms-modeler-api</compileSourcesArtifact>
                         <compileSourcesArtifact>org.jbpm:jbpm-wb-process-runtime-forms-client</compileSourcesArtifact>
-                        <compileSourcesArtifact>org.jbpm:jbpm-wb-human-tasks-forms-modeler-client</compileSourcesArtifact>
+                        <compileSourcesArtifact>org.jbpm:jbpm-wb-forms-modeler-client</compileSourcesArtifact>
                         <compileSourcesArtifact>org.jbpm:jbpm-wb-process-runtime-api</compileSourcesArtifact>
                         <compileSourcesArtifact>org.jbpm:jbpm-wb-process-runtime-client</compileSourcesArtifact>
                         <compileSourcesArtifact>org.jbpm:jbpm-wb-kie-server-api</compileSourcesArtifact>
