+++ jbpm-console-ng-showcase/src/main/webapp/WEB-INF/beans.xml
@@ -17,7 +17,7 @@
     <exclude name="org.jbpm.console.ng.df.client.**"/>
     <exclude name="org.jbpm.console.ng.ht.admin.client.**"/>
     <exclude name="org.jbpm.console.ng.ht.forms.client.**"/>
-    <exclude name="org.jbpm.console.ng.ht.forms.modeler.client.**"/>
+    <exclude name="org.jbpm.workbench.forms.modeler.client.**"/>
     <exclude name="org.jbpm.console.ng.pr.client.**"/>
     <exclude name="org.jbpm.console.ng.gc.forms.client.**"/>
     <exclude name="org.jbpm.console.ng.pr.forms.client.**"/>
