+++ jbpm-wb-forms/jbpm-wb-forms-modeler/jbpm-wb-forms-modeler-api/src/main/java/org/jbpm/workbench/forms/modeler/display/impl/FormModelerFormRenderingSettings.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jbpm.workbench.forms.modeler.display.impl;
+
+import org.jboss.errai.common.client.api.annotations.MapsTo;
+import org.jboss.errai.common.client.api.annotations.Portable;
+import org.jbpm.console.ng.ga.forms.display.FormRenderingSettings;
+
+@Portable
+public class FormModelerFormRenderingSettings implements FormRenderingSettings {
+
+    private String contextId;
+
+    public FormModelerFormRenderingSettings( @MapsTo( "contextId" ) String contextId ) {
+        this.contextId = contextId;
+    }
+
+    public String getContextId() {
+        return contextId;
+    }
+
+    public void setContextId( String contextId ) {
+        this.contextId = contextId;
+    }
+}
