+++ jbpm-wb-forms/jbpm-wb-forms-modeler/jbpm-wb-forms-modeler-api/src/main/resources/org/jbpm/workbench/forms/modeler/JbpmWorkbenchFormsModelerAPI.gwt.xml
