+++ jbpm-wb-forms/jbpm-wb-forms-modeler/jbpm-wb-forms-modeler-client/src/main/resources/org/jbpm/workbench/forms/modeler/JbpmWorkbenchFormsModelerClient.gwt.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+  ~ Copyright 2017 Red Hat, Inc. and/or its affiliates.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~       http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.4.0//EN"
+    "http://google-web-toolkit.googlecode.com/svn/tags/2.5.0/distro-source/core/src/gwt-module.dtd">
+<module>
+
+    <!-- Inherit the core Web Toolkit stuff. -->
+  <inherits name="com.google.gwt.i18n.I18N" />
+  <inherits name="com.google.gwt.http.HTTP" />
+  <inherits name="com.google.gwt.user.User" />
+    
+  <inherits name="org.jboss.errai.ui.UI" />
+  <inherits name="org.uberfire.UberfireClientAPI" />
+  <inherits name="org.jbpm.formModeler.FormModelerApi"/>  
+  <inherits name="org.jbpm.console.ng.ht.forms.JbpmConsoleNGHumanTasksFormsAPI" />
+  <inherits name="org.jbpm.workbench.forms.modeler.JbpmWorkbenchFormsModelerAPI" />
+
+  <inherits name="org.jbpm.console.ng.ht.JbpmConsoleNGHumanTasksAPI" />
+
+  <inherits name="org.kie.workbench.common.widgets.KieWorkbenchWidgetsCommon"/>
+
+</module>
