+++ jbpm-wb-forms/jbpm-wb-forms-modeler/jbpm-wb-forms-modeler-client/src/main/resources/org/jbpm/workbench/forms/modeler/client/i18n/Constants_fr.properties
